Don't try to read zero bytes in `_bulk_reply()`
While investigating transactions blocking indefinitely (or taking minutes to return) I
realized that we only rturn NULL in _bulk_reply() if we detect there are less than
zero bytes to read, but continue if there are non-zero (including zero) bytes to read.

Since the length value passed to fgets() is padded by one (+1), we perenially block
for a byte that just does not exist in the stream in that case.

This trivial patch fixes that.

Signed-off-by: Tim Post <tpost@ezp.net>